bpf: Fix compile-tested macros used in is_defined
In the datapath, we have two ways to check if a macro is defined:
the usual defined()/#ifdef and is_defined(). The later is often used in
C predicates or in tail call conditions. It expects the macro to be
defined with a value, typically #define ENABLE_FEATURE_X 1. Without this
non-zero value, is_defined(ENABLE_FEATURE_X) returns false.

This commit fixes our compilation tests to set a value for all macros
that are used in is_defined. The list was extracted with the following
command:

    git grep -hoP "(?<=is_defined\()[_A-Z0-9]+" bpf/ | sort | uniq

Signed-off-by: Paul Chaignon <paul@cilium.io>