test: Un-Quarantine K8sUpdates on GKE
[ upstream commit 699ca469c90f80d9352c8019fc88f640bbf83370 ]

This removes the quarantine check for the K8sUpdates test on GKE. The
issues with upgrade tests had on GKE were addressed in #14187, #14294,
 #14295, and #14393.

The last fix was merged 6 days ago. The quarantined GKE pipeline had
since no failures that look related to the issues with DNS that we
previously observed.

Fixes: #13833
Fixes: #13898

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Chris Tarazi <chris@isovalent.com>